% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cutLD}
\alias{cutLD}
\title{cut the LD with new thresh}
\usage{
cutLD(ldm, outfile, cutThresh = 1)
}
\arguments{
\item{ldm}{string, input ldm}

\item{outfile}{string, output file}
}
\description{
cut the LD with new thresh
}
