% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutLDs.R
\name{cutLDs}
\alias{cutLDs}
\title{cut the LD to a different variance threshold to save the storage}
\usage{
cutLDs(inputDir, outputDir, 0.999)
}
\arguments{
\item{inputDir}{string, LD path for input}

\item{outputDir}{string, LD path for output}

\item{thresh}{numeric, new threshold to cut}
}
\value{
none, the output is in the output file
}
\description{
cut the LD to a different variance threshold to save the storage
}
