% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_eig.r
\name{impute}
\alias{impute}
\title{Impute summary data}
\usage{
impute(mafile, LD_folder, output)
}
\arguments{
\item{mafile}{string, the path of summary data in COJO format}

\item{LD_folder}{string,  path of LD folder}

\item{output}{string, path to output}

\item{thresh}{number, eigen cutoff threshold for imputation, default 0.995}
}
\value{
none, the output is in the output file
}
\description{
Impute summary data
}
