% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LDstep1}
\alias{LDstep1}
\title{LDstep1: generate the LD command line}
\usage{
LDstep1(mafile, genoPrefix, outDir)
}
\arguments{
\item{mafile}{string, path to the summary file}

\item{genoPrefix}{string, path to the genotype, if multiple genotype files, use {CHR} to indicate the chromosome}

\item{outDir}{string, folder to save the output (the function will create the folder)}

\item{genoCHR}{string, if multiple genotype file present, change this to "startCHR-endCHR", e.g., 1-22,X: means from chr1 to chr22, and chrX}

\item{blockRef}{string, reference window file downloaded from our website or customized, the default 4cM window (in GRch37) will be provided if empty}

\item{tool}{string, the command line to call the underlying tool (currently gctb), make sure the tool can be run in terminal directly}

\item{log2file}{boolean, FALSE: display message on terminal (default); TRUE: redirect to an output file}
}
\value{
none, results in the specified output
}
\description{
LDstep1: generate the LD command line
}
