% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LDstep2}
\alias{LDstep2}
\title{LDstep2: run LD}
\usage{
LDstep2(outDir, blockIndex)
}
\arguments{
\item{outDir}{string, path to LD folder}

\item{blockIndex}{numeric, block index to run}

\item{log2file}{boolean, FALSE: display message on terminal; TRUE: redirect to an output file; default FALSE}
}
\value{
none, results in the specified output
}
\description{
LDstep2: run LD
}
