% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LDstep3}
\alias{LDstep3}
\title{LDstep3: eigen decomposition on LD matrix}
\usage{
LDstep3(outDir, blockIndex)
}
\arguments{
\item{outDir}{string, path to the LD folder}

\item{blockIndex}{numeric, perform analysis on a specified block number}

\item{thresh}{numeric, variance cutoff threshold, suggested default 0.995}

\item{log2file}{boolean, FALSE: display message on terminal (default); TRUE: redirect to an output file}
}
\value{
none, results in the specified output
}
\description{
LDstep3: eigen decomposition on LD matrix
}
