% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{LDstep4}
\alias{LDstep4}
\title{LDstep4: Merge the LD information}
\usage{
LDstep4(outDir)
}
\arguments{
\item{outDir}{string, path to the LD folder}

\item{log2file}{boolean, FALSE: display message on terminal; TRUE: redirect to an output file; default value is FALSE}
}
\value{
none, results in the specified output
}
\description{
LDstep4: Merge the LD information
}
