% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbr_eigen.r
\name{sbayesrc}
\alias{sbayesrc}
\title{SBayesRC with eigen decomposition}
\usage{
sbayesrc(ma_path, LD_folder, file_out)
}
\arguments{
\item{file_summary}{string, summary data path, COJO format}

\item{ld_folder}{string, path to LD folder}

\item{file_out}{string, output path}

\item{thresh}{number, eigen cutoff for LD}

\item{niter}{number, total number of MCMC iterations}

\item{burn}{number, numbmer of MCMC burn-in iterations}

\item{fileAnnot}{sring, annotation file path}

\item{nComp}{number, number of mixture component}

\item{starth2}{number, start heritability}

\item{startPi}{vector, start proportion of causal in each component}

\item{gamma}{vector, gamma of each componnet}
}
\value{
none, output to the file_out
}
\description{
SBayesRC with eigen decomposition
}
\examples{
# run SBayesRC without annotation
sbayesrc(ma_path, LD_folder, file_out)
# run SBayesRC with annotation
sbayesrc(ma_path, LD_folder, file_out, fileAnnot=annot_file_path)
}
