% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{sbrcMulti}
\alias{sbrcMulti}
\title{caculate SBayesRC-multi PRS}
\usage{
sbrcMulti(prs1, prs2, outPrefix, tuneid, pheno)
}
\arguments{
\item{prs1}{string, path to the polygenic score set1 from SBayesRC::prs}

\item{prs2}{string, path to the polygenic score set2 from SBayesRC::prs}

\item{outPrefix}{string, output path}

\item{tuneid}{string, path to sample id (FID IID) for the tunning}

\item{pheno}{string, path to the phenotype (FID IID pheno; without header)}

\item{keepid}{string, path to sample id (FID IID) to be kept, default "": all samples}

\item{log2file}{boolean, FALSE: display message on terminal; TRUE: redirect to an output file; default value is FALSE}
}
\value{
none, results in the specified output
}
\description{
caculate SBayesRC-multi PRS
}
