% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ma.R
\name{tidy}
\alias{tidy}
\title{Tidy GWAS summary data}
\usage{
tidy(ma_path, LD_folder, output)
}
\arguments{
\item{ma_file}{string, summary data path, COJO format}

\item{ld_folder}{string,  path to LD folder}

\item{output}{string, output path}

\item{freq_thresh}{number, max difference in allele frequency}

\item{N_sd_range, }{number, filter the per SNP sample size in mean +- N_sd_range * sd}
}
\value{
none, the output is in the output file
}
\description{
Tidy GWAS summary data
}
