% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutLDs.R
\name{cutLD}
\alias{cutLD}
\title{cut the LD to a different variance threshold to save the storage}
\usage{
cutLD(inputDir, outDir)
}
\arguments{
\item{inputDir}{string, path for input LD}

\item{outDir}{string, path for output LD}

\item{thresh}{numeric, new threshold to cut, default 0.995}

\item{log2file}{boolean, FALSE: display message on terminal; TRUE: redirect to an output file; default FALSE}
}
\value{
none, results in the specified output
}
\description{
cut the LD to a different variance threshold to save the storage
}
