% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_eig.r
\name{impute}
\alias{impute}
\title{Impute summary data}
\usage{
impute(mafile, LDdir, output)
}
\arguments{
\item{mafile}{string, the path of summary data in COJO format}

\item{LDdir}{string,  path of LD folder}

\item{output}{string, path to output (with filename.ma)}

\item{thresh}{number, eigen cutoff threshold for imputation, default 0.995}

\item{log2file}{boolean, FALSE: display message on terminal; TRUE: redirect to an output file; default FALSE}
}
\value{
none, results in the specified output
}
\description{
Impute summary data
}
