% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBayesRC-package.R
\docType{package}
\name{SBayesRC-package}
\alias{SBayesRC}
\alias{SBayesRC-package}
\title{SBayesRC: SBayesRC Method for Polygenic Prediction}
\description{
This package implemented SBayesRC that incorporates functional annotations with high-density variants for polygenic prediction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zhilizheng/SBayesRC}
  \item Report bugs at \url{https://github.com/zhilizheng/SBayesRC/issues}
}

}
\author{
\strong{Maintainer}: Zhili Zheng \email{zhilizheng@outlook.com}

Authors:
\itemize{
  \item Jian Zeng \email{j.zeng@uq.edu.au}
}

}
\keyword{internal}
