% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbr_eigen.r
\name{sbayesrc}
\alias{sbayesrc}
\title{SBayesRC analysis}
\usage{
sbayesrc(mafile, LDdir, outPrefix, annot="AnnotFile_PATH")
}
\arguments{
\item{mafile}{string, path of summary data in COJO format}

\item{LDdir}{string, path to LD folder}

\item{outPrefix}{string, output path}

\item{annot}{string, annotation file path}

\item{log2file}{boolean, FALSE: display message on terminal; TRUE: redirect to an output file; the default is FALSE}

\item{bTune}{boolean, perform tuning or not (TRUE, FALSE), default TRUE (yes)}

\item{tuneIter}{integer, total number of tuning iterations}

\item{tuneBurn}{integer, number of tuning iterations to burn}

\item{thresh}{numeric, eigen variance cutoff for LD, default 0.995}

\item{tuneStep}{numeric vector, eigen variance cutoffs to search}

\item{bTunePrior}{boolean, use prior estimated from tuning step or not}

\item{niter}{integer, total number of MCMC iterations}

\item{burn}{integer, numbmer of MCMC burn-in iterations}

\item{starth2}{numeric, heritability value to start iteration}

\item{startPi}{vector, start proportion of causal in each component}

\item{gamma}{vector, gamma of each componnet}

\item{method}{string, method to use (always sbr_ori)}

\item{sSamVe}{string, method to re-sample residual (always allMixVe)}

\item{twopq}{string, method to scale the beta}

\item{bOutDetail}{boolean, output more details or not}

\item{resam_thresh}{numeric, threshold to resampling the residual variance}

\item{seed}{integer, random seed}

\item{outFreq}{integer, frequencies to calculate the parameters and output}

\item{annoSigmaScale}{numeric, scale to the annotation Sigma}

\item{bOutBeta}{boolean, output raw beta or not}

\item{exclude}{string, path to a file contains the SNPs excluded from modeling, defalut empty;}
}
\value{
none, results in the specified output
}
\description{
SBayesRC analysis
}
\examples{
## run SBayesRC without annotation
#sbayesrc(mafile, LDdir, output)
## run SBayesRC with annotation
# sbayesrc(mafile, LDdir, ouput, annot=annot_file_path)
}
