% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ma.R
\name{tidy}
\alias{tidy}
\title{Tidy GWAS summary data}
\usage{
tidy(mafile, LDdir, output)
}
\arguments{
\item{mafile}{string, summary data path, COJO format}

\item{LDdir}{string,  path to LD folder}

\item{output}{string, output path}

\item{freq_thresh}{numeric, max difference in allele frequency}

\item{N_sd_range}{numeric, filter the per SNP sample size in mean +- N_sd_range * sd}

\item{rate2pq}{numeric, check ratio of variance estimated by 2 methods in summary data}

\item{log2file}{boolean, FALSE: display message on terminal; TRUE: redirect to an output file; default value is FALSE}
}
\value{
none, results in the specified output
}
\description{
Tidy GWAS summary data
}
